django-app/pyproject.toml
[tool.poetry]
name = "harsh-realm"
version = "0.1.0"
description = ""
authors = ["Josh Madden <cyrex562@gmail.com>"]
readme = "README.md"
packages = [{include = "harsh_realm"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-import-export = "^3.2.0"
django-material-admin = "^1.8.6"
django-htmx = "^1.14.0"
django-debug-toolbar = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
